requirements.txt
import os
# Data Collection APIs
ccxt==4.1.97
pycoingecko==3.1.0
web3==6.12.0
requests==2.31.0
yfinance==0.2.38

# Data Manipulation
pandas==2.2.2
numpy==1.26.4


# Data Visualization
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.20.0
dash==2.12.0  # for interactive web-based visualizations
dash-bootstrap-components==2.0.1


# Statistical and Financial Analysis
#scipy==1.13.0
statsmodels==0.14.1
empyrical==0.5.5
pyfolio==0.9.2  # (may require patching for latest pandas versions)

# Other Useful Tools
python-dotenv==1.0.1  # for managing API keys securely
tqdm==4.66.2  # progress bars
jupyterlab==4.1.6  # if using notebooks
pylint==2.20.0  # for code quality checks
# Optional: GraphQL query tools (for The Graph)
gql==3.5.0
aiohttp==3.9.3
